eld_logs/pyproject.toml
[project]
name = "eld-logs"
version = "0.1.0"
description = "ELD logs and route planning API"
readme = "README.md"
requires-python = ">=3.14"

dependencies = [
    "Django==6.0.0",
    "djangorestframework==3.16.1",
    "django-cors-headers==4.9.0",
    "celery==5.6.0",
    "redis==7.1.0",
    "openrouteservice==2.3.3",
    "Pillow==12.0.0",
    "python-dotenv==1.2.1",
    "requests==2.32.5",
    "python-decouple==3.8",
    "drf-spectacular==0.29.0",
    "staticmap==0.5.7",
    "psycopg[binary]>=3.3.2",
    "channels>=4.3.2",
    "channels-redis>=4.3.0",
    "uvicorn>=0.38.0",
    "django-celery-results>=2.6.0",
    "websockets>=15.0.1",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
    "pytest-cov>=7.0.0",
    "pytest-asyncio>=1.3.0",
    "black>=25.12.0",
    "ruff>=0.1.0",
    "mypy>=1.19.1",
    "django-stubs>=5.2.8",
    "djangorestframework-stubs>=3.16.6",
    "factory-boy==3.3.3",
    "pytest-mock>=3.15.1",
    "daphne>=1.7.0",
]

prod = [
    "gunicorn>=23.0.0",
]

[tool.uv]
compile-bytecode = true
resolution = "highest"

[tool.mypy]
python_version = "3.14"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
strict = true

[tool.django-stubs]
django_settings_module = "eld_logs.settings.local"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "eld_logs.settings.test"
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
